+++ tests/shapes/test_solid.py
@@ -1,8 +1,13 @@
-import math
+import os
 import unittest
+import math
+from dessia_common.core import DessiaObject
 
 import volmdlr
 from volmdlr import curves, faces, shapes, surfaces, wires, primitives3d
+
+folder = os.path.join(os.path.dirname(os.path.realpath(__file__)))
+objects_folder = os.path.join(os.path.dirname(os.path.realpath(__file__)), "shapes_objects")
 
 
 class TestSolid(unittest.TestCase):
@@ -20,6 +25,19 @@ class TestSolid(unittest.TestCase):
     faces2 = [f.translation(volmdlr.Vector3D(1, 1, 1)) for f in faces_list]
     solid2 = shapes.Solid.make_solid(shapes.Shell(faces=faces2))
 
+    def test_check_platform(self):
+        self.assertIsNone(self.solid1._check_platform())
+
+    def test_to_dict_dict_to_object(self):
+        to_dict = self.solid1.to_dict()
+        dict_to_obejct = DessiaObject.dict_to_object(to_dict)
+        self.assertEqual(dict_to_obejct, self.solid1)
+
+    def test_to_brep_from_brep(self):
+        self.solid1.to_brep(objects_folder+"/test_to_brep.brep")
+        from_brep = shapes.Solid.from_brep(objects_folder+"/test_to_brep.brep")
+        self.assertEqual(from_brep, self.solid1)
+
     def test_union(self):
         union = self.solid1.union(self.solid2)
         self.assertAlmostEqual(union.volume(), 15)
@@ -31,6 +49,28 @@ class TestSolid(unittest.TestCase):
     def test_intersection(self):
         intersection = self.solid1.intersection(self.solid2)
         self.assertAlmostEqual(intersection.volume(), 1.0)
+
+    def test_box(self):
+        box = shapes.Solid.make_box(length=2, width=3, height=5)
+        self.assertEqual(box.volume(), 2 * 3 * 5)
+
+    def test_cone(self):
+        cone = shapes.Solid.make_cone(radius1=0, radius2=5, height=5, direction=volmdlr.X3D, angle_degrees=270)
+        self.assertAlmostEqual(cone.volume(), 98.1747704332957)
+
+    def test_cylinder(self):
+        cylinder = shapes.Solid.make_cylinder(radius=5, height=5, direction=volmdlr.X3D, angle_degrees=270)
+        self.assertAlmostEqual(cylinder.volume(), 294.5243112989299)
+
+    def test_sphere(self):
+        sphere1 = shapes.Solid.make_sphere(radius=5, direction=volmdlr.X3D, angle_degrees1=-90,
+                                           angle_degrees2=60, angle_degrees3=270)
+        self.assertAlmostEqual(sphere1.volume(), 387.6486925590814)
+
+    def test_torus(self):
+        torus1 = shapes.Solid.make_torus(radius1=2, radius2=.5, direction=volmdlr.X3D, angle_degrees1=0,
+                                         angle_degrees2=360)
+        self.assertAlmostEqual(torus1.volume(), 9.869604401089358)
 
     def test_make_extrusion(self):
         length, width, height, radius = 0.4, 0.3, 0.08, 0.1
@@ -61,20 +101,22 @@ class TestSolid(unittest.TestCase):
 
     def test_sweep(self):
         point1, point2 = volmdlr.Point3D(1.0, 1.0, 0.0), volmdlr.Point3D(1.0, 0.5, 0.0)
-        path = primitives3d.OpenRoundedLineSegments3D([point1, point2, volmdlr.Point3D(0.5, 0.5, 0.0)], {"1": 0.2})
-        section = wires.Contour2D.from_circle(curves.Circle2D(volmdlr.OXY, 0.05))
-        # section = section.to_3d(volmdlr.O3D, volmdlr.X3D, volmdlr.Y3D)
+        point3 = volmdlr.Point3D(0.5, 0.5, 0.0)
+        path = primitives3d.OpenRoundedLineSegments3D([point1, point2, point3], {"1": 0.2})
+        outer_contour = wires.Contour2D.from_circle(curves.Circle2D(volmdlr.OXY, 0.05))
+        inner_contours = [wires.Contour2D.from_circle(curves.Circle2D(volmdlr.OXY, 0.045))]
         direction = (point2 - point1).unit_vector()
         frame = volmdlr.Frame3D.from_point_and_vector(point=point1, vector=direction, main_axis=volmdlr.Z3D)
         section = faces.PlaneFace3D(surface3d=surfaces.Plane3D(frame=frame),
-                                    surface2d=surfaces.Surface2D(outer_contour=section, inner_contours=[]))
-        sweep = shapes.Solid.make_sweep(section, path)
-        self.assertEqual(len(sweep.faces), 5)
-        for face, expected_face_class in zip(sweep.faces, [faces.PlaneFace3D, faces.CylindricalFace3D,
-                                                           faces.ToroidalFace3D, faces.CylindricalFace3D,
-                                                           faces.PlaneFace3D]):
-            self.assertTrue(isinstance(face, expected_face_class))
+                                    surface2d=surfaces.Surface2D(outer_contour=outer_contour, inner_contours=inner_contours))
+        sweep = shapes.Solid.make_sweep_from_contour(section, path, inner_contours)
+        self.assertEqual(len(sweep.primitives[0].primitives), 8)
 
+        path = wires.Wire3D.from_points([point1, point2, point3])
+        sweep1 = shapes.Solid.make_sweep(face=section, path=path, transition_mode="right")
+        self.assertEqual(len(sweep1.primitives[0].primitives), 6)
+        sweep2 = shapes.Solid.make_sweep(face=section, path=path, transition_mode="round")
+        self.assertEqual(len(sweep2.primitives[0].primitives), 10)
 
 if __name__ == '__main__':
     unittest.main()