+++ CHANGELOG.md
@@ -14,9 +14,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Shape: volume, bounding_box
 - Shell: add 3 different instanciation methods
 - Shape: add boolean operations: subtraction, union, intersection
+<<<<<<< HEAD
 - Shell constructors: make_extrusion, make_wedge
 - Solid constructors: make_extrusion, make_extrusion_from_frame_and_wires, make_wedge, make_sweep
 
+=======
+- Shape: to_dict, dict_to_object, to_brep, from_brep, from_brep_stream, to_brep_stream, 
+- Solid: make_box, make_cone, make_cylinder, make_torus, make_sphere methods.
+>>>>>>> origin/solid_geometries
 
 #### core.py
 - VolumeModel: from_step using OCP step reader